Added freeApartments under freeData.c
