+++ Final_Project/Final_Project.vcxproj
@@ -151,6 +151,7 @@
     <ClCompile Include="buyAnApt.c" />
     <ClCompile Include="commander.c" />
     <ClCompile Include="deleteAnApt.c" />
+    <ClCompile Include="freeData.c" />
     <ClCompile Include="getAnApt.c" />
     <ClCompile Include="main.c" />
     <ClCompile Include="ReadDataFromFiles.c" />
