+++ Final_Project/archiveManager.c
@@ -198,34 +198,8 @@ void archivePrinter(History_Data * hData, int task) {
 					break;
 			}
 		}
-	}
+	} 
 	else {
 		return;
-	}
-}
-
-void freeHistoryData(History_Data * hData) {
-	freeShortTermHisArr(hData->shortTerm_HistoryArr);
-	freeHistoryList(&hData->LongTerm_HistoryList);
-	return;
-}
-
-void freeShortTermHisArr(char ** shortTermArr) {
-	int i;
-	for (i = SHORT_TERM_SIZE - 1; i >= 0; i--) {
-		if (shortTermArr[i] != NULL)
-			free(shortTermArr[i]);
-	}
-}
-
-void freeHistoryList(HistoryList * hList) {
-	HistoryEntry * p = hList->tail;
-	HistoryEntry *  temp;
-	while (p != NULL) {
-		if (p->command != NULL)
-			free(p->command);
-		temp = p;
-		p = p->prev;
-		free(temp);
 	}
 }
